requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.1
azure-functions==1.12.0       
azure-functions-durable==1.2.1
azure-storage-blob==12.14.1   
beautifulsoup4==4.11.1        
bootstrap3-datetime==2.4      
bs4==0.0.1
certifi
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
Django==4.1.4
django-environ==0.9.0
django-flatpickr==2.0.0
fonttools==4.38.0
frozenlist==1.3.3
furl==2.1.3
html5lib==1.1
idna==3.4
isodate==0.6.1
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
msrest==0.7.1
multidict==6.0.3
numpy==1.23.5
oauthlib==3.2.2
orderedmultidict==1.0.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
plotly==5.11.0
psutil==5.9.4
pycparser==2.21
pydantic==1.10.4
pyodbc==4.0.35
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==8.1.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
webencodings==0.5.1
wincertstore==0.2
yarl==1.8.2